+++ CHANGELOG.rst
@@ -5,6 +5,18 @@ Changelog
 -------------
 * Added ``dms_subassemble`` program
 
+1.1.11
+-------------
+* Fixed bug in ``dms_merge``, where normalization of prefs or diffprefs was not occurring when the user provided two or more identical pref or diffpref files
+
+* Added y-axis label override option in ``plot.PlotDepth()`` and corrected ``dms_summarizealignments`` to label the y-axis as 'number of barcodes' when summarizing barcoded subamplicon data.
+
+* Added ``merge_method`` choice ``rescale`` to ``dms_merge`` to rescale a preferences file by a ``--stringencyparameter``.
+
+1.1.9
+-------------
+* Added option to trim reads R1 and R2 in ``dms_barcodedsubamplicons``
+
 1.1.8
 --------
 * Fixed algorithm to estimate preferences and differential preferences using the ``--ratio_estimation`` option for both ``dms_inferprefs`` and ``dms_inferdiffprefs``.